move render time out of the main component
